requirements.txt
boto3==1.14.15
Flask==1.1.2
line-bot-sdk==1.16.0
requests==2.24.0
